Fix Android 12 can't start and stop cfm

+++ app/src/main/assets/CFM_Start.sh
@@ -5,5 +5,9 @@ else
   MAGISK_PATH="/data/adb/modules"
 fi
 
-disABpath=${MAGISK_PATH}/Clash_For_Magisk/disable
-rm -f ${disABpath}
+if [ "$(getprop ro.build.version.release)" -lt "12" ] ; then
+  disABpath=${MAGISK_PATH}/Clash_For_Magisk/disable
+  rm -f ${disABpath}
+else
+  ${MAGISK_PATH}/Clash_For_Magisk/scripts/clash.service -s && ${MAGISK_PATH}/Clash_For_Magisk/scripts/clash.tproxy -s
+fi
